Tunings tests and refactor
