+++ src/xtgeo/grid3d/_grid_import_roff.py
@@ -4,7 +4,6 @@
 from __future__ import print_function, absolute_import
 
 from collections import OrderedDict
-import json
 
 import numpy as np
 
@@ -190,16 +189,7 @@ def _import_roff_xtgformat2(self, gfile):
     logger.info("Initilize arrays... done")
 
     _rkwxvec_coordsv(
-        self,
-        gfile,
-        kwords,
-        byteswap,
-        xshift,
-        yshift,
-        zshift,
-        xscale,
-        yscale,
-        zscale,
+        self, gfile, kwords, byteswap, xshift, yshift, zshift, xscale, yscale, zscale,
     )
 
     logger.info("ZCORN related...")
@@ -223,12 +213,7 @@ def _import_roff_xtgformat2(self, gfile):
 
     logger.info("ACTNUM...")
     self._actnumsv = _rkwxvec_prop(
-        self,
-        gfile,
-        kwords,
-        "active!data",
-        byteswap,
-        strict=False,
+        self, gfile, kwords, "active!data", byteswap, strict=False,
     )
     if self._actnumsv is None:
         self._actnumsv = np.ones((self._ncol, self._nrow, self._nlay), dtype=np.int32)
@@ -236,49 +221,3 @@ def _import_roff_xtgformat2(self, gfile):
     logger.info("ACTNUM... done")
     logger.info("XTGFORMAT is %s", self._xtgformat)
 
-
-def import_xtgeo(self, gfile):
-    """Import xtgeo format (in prep)"""
-
-    fhandle = gfile.get_cfhandle()
-
-    # scan dimensions:
-    ier, ncol, nrow, nlay, _metadata = _cxtgeo.grdcp3d_import_xtgeo_grid(
-        0,
-        np.zeros((0), dtype=np.float64),
-        np.zeros((0), dtype=np.float32),
-        np.zeros((0), dtype=np.int32),
-        fhandle,
-    )
-
-    if ier != 0:
-        raise ValueError("Cannot import, error code is: {}".format(ier))
-
-    self._ncol = ncol
-    self._nrow = nrow
-    self._nlay = nlay
-
-    nncol = ncol + 1
-    nnrow = nrow + 1
-    nnlay = nlay + 1
-
-    self._coordsv = np.ones((nncol, nnrow, 6), dtype=np.float64)
-    self._zcornsv = np.zeros((nncol, nnrow, nnlay, 4), dtype=np.float32)
-    self._actnumsv = np.zeros((ncol, nrow, nlay), dtype=np.int32)
-
-    # read data
-    ier, ncol, nrow, nlay, _metadata = _cxtgeo.grdcp3d_import_xtgeo_grid(
-        1,
-        self._coordsv,
-        self._zcornsv,
-        self._actnumsv,
-        fhandle,
-    )
-
-    if ier != 0:
-        raise ValueError("Error in reading file, error code is: {}".format(ier))
-
-    # themeta = json.loads(metadata, object_pairs_hook=OrderedDict)
-    self._subgrids = None  # tmp
-
-    gfile.cfclose()
